feature: authentication with Github and Twitter
